% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVG.R
\docType{class}
\name{SVG-class}
\alias{SVG}
\alias{SVG-class}
\title{Make svg for collocations graph}
\description{
The resulting svg graph will have clickable nodes and edges. To use this functionality
you need to install FastRWeb and start it with sudo /var/FastRWeb/code/start
}
\section{Fields}{

\describe{
\item{\code{edgeAttributes}}{attributes of edges to maintain}

\item{\code{verticeAttributes}}{attributes of vertices to maintain}

\item{\code{as.undirected}}{logical, whether to turn object into directed graph}

\item{\code{object}}{the collocation object}

\item{\code{layout}}{either "kamada.kawai" or "fruchterman.reingold"}

\item{\code{width}}{the width of the svg}

\item{\code{height}}{the height of the svg}

\item{\code{margin}}{margins of the svg}

\item{\code{fontSize}}{font size of the vertex labels}

\item{\code{textOffset}}{where to put text}

\item{\code{edgeAttributes}}{attributes of edges for tooltips}

\item{\code{verticeAttributes}}{attributes of attributes for tooltips}

\item{\code{pandocTab}}{logical, whether to format tables with pandoc}
}}
\section{Methods}{

\describe{
\item{\code{as.igraph(as.undirected = TRUE)}}{Turn SVG into igraph object.}

\item{\code{browse()}}{Show SVG in browser. The xml/svg file is stored in a temporary file, which is
returned invisibly.}

\item{\code{html()}}{Turn into html.}

\item{\code{initialize(x)}}{Create new SVG object.}

\item{\code{make(verbose = TRUE)}}{Create XML.}

\item{\code{makeEdges()}}{Make edges.}

\item{\code{makeNodes(radius = list(minSize = 5, tf = TRUE), pandocTab = TRUE)}}{Make nodes.}

\item{\code{plot()}}{Show SVG in viewer pane of RStudio.}

\item{\code{selectCommunity(community)}}{Select a community.}

\item{\code{store(filename = tempfile(fileext = ".html"))}}{Save XML/SVG to disk.}

\item{\code{textNodes(fontSize = 8, textOffset = 3)}}{Make text nodes.}
}}
\examples{
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_speaker="Angela Merkel", text_type="speech"), tf="word")
}
}

