% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.three_method.R
\docType{methods}
\name{as.three}
\alias{as.three}
\alias{as.three,igraph-method}
\alias{as.three,list-method}
\title{generate 3d graph}
\usage{
as.three(object, ...)

\S4method{as.three}{igraph}(object, type = "base", bgColor = "0xffffff",
  nodeSize = 5, edgeColor = "0x000000", edgeWidth = 5, fontSize = 20,
  fontColor = "0x000000", fontOffset = c(x = 5, y = 5, z = 5),
  jsUrlPrefix = NULL, adjust = list())

\S4method{as.three}{list}(object, type = "base", bgColor = "0xffffff",
  nodeSize = 5, edgeColor = "0x000000", edgeWidth = 5, fontSize = 20,
  fontColor = "0x000000", fontOffset = c(x = 5, y = 5, z = 5),
  jsUrlPrefix = NULL, adjust = list())
}
\arguments{
\item{object}{the object}

\item{type}{display mode}

\item{bgColor}{a hex value}

\item{nodeSize}{size of nods}

\item{edgeColor}{a hex value}

\item{edgeWidth}{defaults to 5}

\item{fontSize}{defaults to 20}

\item{fontColor}{a hex value}

\item{fontOffset}{a list}
}
\description{
generate 3d graph with three.js
}

