% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collocations.R, R/keynessCollocations.R, R/svg.R
\docType{methods}
\name{asIgraph,collocations-method}
\alias{asIgraph}
\alias{asIgraph,collocations-method}
\alias{asIgraph,keynessCollocations-method}
\alias{asIgraph,svg-method}
\title{Turn collocations object into igraph}
\usage{
\S4method{asIgraph}{collocations}(x, edgeAttributes = "ll",
  verticeAttributes = "tf", as.undirected = TRUE)

\S4method{asIgraph}{keynessCollocations}(x)

\S4method{asIgraph}{svg}(x, as.undirected = TRUE)
}
\arguments{
\item{x}{an object}

\item{edgeAttributes}{attributes of edges to maintain}

\item{verticeAttributes}{attributes of vertices to maintain}

\item{as.undirected}{logical, whether to turn object into directed graph}
}
\description{
Turn collocations object into igraph

Turn a keyness collocations object into an igraph-object
}
\examples{
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_speaker="Angela Merkel"))
bt17cducsu <- partition("PLPRTXT", list(text_lp="17", text_type="speech", text_party="CDU_CSU"))
bt17merkelColl <- collocations(bt17merkel, pAttribute="lemma", progress=TRUE)
bt17cducsuColl <- collocations(bt17cducsu, pAttribute="lemma", progress=TRUE)
merkel <- keyness(bt17merkelColl, bt17cducsuColl, included=TRUE, mc=FALSE)
merkelG <- asIgraph(merkel)
}
}

