% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurrences.R
\docType{class}
\name{Cooccurrences-class}
\alias{Cooccurrences}
\alias{Cooccurrences-class}
\title{Get cooccurrences.}
\description{
Get cooccurrences.
}
\section{Fields}{

\describe{
\item{\code{partition}}{partition object}

\item{\code{pAttribute}}{character}

\item{\code{drop}}{character vector}

\item{\code{method}}{statistical test ("ll")}

\item{\code{window}}{integer}

\item{\code{verbose}}{logical}

\item{\code{contextSizes}}{data.table}

\item{\code{dt}}{data.table}
}}
\examples{
\dontrun{
merkel_partition <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_lp = "17", speaker_type = "speech", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(merkel_partition, pAttribute = "word"))))
Merkel <- Cooccurrences$new(partition = merkel_partition, pAttribute = "word", window = 5L, drop = termsToDrop)
Merkel$getCounts()
Merkel$trim(drop = termsToDrop)
Merkel$getStats()
tcm <- Merkel$makeTermCooccurrenceMatrix()

bt13_partition <- partition("PLPRBT", speaker_lp = "13", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(.Object, pAttribute = "word"))))
system.time(bt13 <- cooccurrences2(bt13_partition, drop = termsToDrop, tcm = FALSE, method = "ll"))

kohl_partition <- partition("PLPRBT", speaker_name = "Helmut Kohl", speaker_type = "speech", speaker_lp = "13", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(kohl_partition, pAttribute = "word"))))
kohl <- cooccurrences2(kohl_partition, drop = termsToDrop, tcm = FALSE, method = "ll")

kohl_compared <- compare(kohl, bt13, included = TRUE)
kohl_trimmed <- trim(kohl, by = subset(kohl_compared, rank_ll <= 250))

library(polmineR.graph)
library(igraph)
K <- asIgraph(kohl_trimmed)
K <- enrich(K, community = list(method = "fastgreedy", weights=FALSE))
K <- enrich(K, layout = "kamada.kawai", dim = 3)
K <- three::rescale(K, -400, 400)
t <- as.three(K, bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=3, fontOffset=c(x=10,y=10,z=10))
t <- as.three(K, type="anaglyph", bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=3, fontOffset=c(x=10,y=10,z=10))
toView <- three:::store(t, directory="/Users/blaette/Lab/tmp/three")
shiny::runApp("/Users/blaette/Lab/github/polmineR.graph/inst/three")
browseURL(toView["tmpFileJs"])
}
}

