% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ego.R
\docType{class}
\name{Ego-class}
\alias{Ego}
\alias{Ego-class}
\title{Analyse ego networks of queries.}
\arguments{
\item{coi}{bla}

\item{query}{bla}

\item{left}{bla}

\item{right}{bla}

\item{pAttribute}{bla}

\item{min.count}{bla}

\item{min.ll}{bla}

\item{max.coocs}{= 10 bla}

\item{keep}{bla}

\item{drop}{bla}

\item{verbose}{bla}
}
\description{
Analyse ego networks of queries.
}
\section{Fields}{

\describe{
\item{\code{query}}{bla}

\item{\code{left}}{bla}

\item{\code{right}}{bla}

\item{\code{partition}}{bla}

\item{\code{Corpus}}{bla}

\item{\code{pAttribute}}{bla}

\item{\code{keep}}{bla}

\item{\code{drop}}{bla}

\item{\code{min.count}}{bla}

\item{\code{min.ll}}{bla}

\item{\code{max.coocs}}{bla}

\item{\code{igraph}}{bla}

\item{\code{fontSize}}{bla}

\item{\code{layout}}{bla}

\item{\code{verbose}}{bla}
}}
\section{Methods}{

\describe{
\item{\code{initialize(coi, query, left = getOption("polmineR.left"),
  right = getOption("polmineR.right"),
  pAttribute = getOption("polmineR.pAttribute"), min.count = 3,
  min.ll = 10.83, max.coocs = 10, keep = list(), drop = list(word =
  c("\\"", "!", "/", ".", "[", "]")), verbose = TRUE)}}{Create new Ego object.}

\item{\code{plot(with = "networkD3")}}{Show SVG in viewer pane of RStudio.}
}}
\examples{
\dontrun{
E <- Ego$new(
  coi = partition("PLPRBT", text_year = "2010"), query = "Ungleichheit",
  left = 10, right = 10,
  drop = list(
    word = c('"', '!', '/', '.', ":", "[", "]", "#", "'", ","),
    pos = c("ART", "APPR", "PRELS", "PDAT", "$(", "$.", "$,", "APPRAR",
             "VAFIN", "APPRART", "KOUI", "KON", "PPOSAT", "KOUS")
   )
)
E$calc()

E$plot(with = "networkD3")

graph <- as.dgr_graph(E$igraph)
render_graph(graph)
E$plot(with = "DigrammeR")

S <- E$as.SVG()
S$width <- 500
S$height <- 500
S$fontSize <- 12
S$make()
S$browse()


}
}

