% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradget.R
\name{gradget}
\alias{gradget}
\title{Graph Annotation with Shiny.}
\usage{
gradget(graph)
}
\arguments{
\item{graph}{An \code{igraph} object that is enhanced for visualisation with
the gradget htmlwidget (x, y, and z coordinates, if applicable; additional
edge and node data).}
}
\description{
A shiny gadget for graph annotation with shiny, or granny in short.
}
\examples{
library(magrittr)
library(polmineR)
library(pbapply)
use("GermaParl")

G <- merkel2008 \%>\%
  igraph_add_coordinates(layout = "kamada.kawai", dim = 3) \%>\%
  igraph_add_communities() \%>\% 
  rescale(-250, 250)

am2008 <- partition(
  "GERMAPARL",
  speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word"
)
# G <- igraph_add_kwic(G, subcorpus = am2008)
if (interactive()) gradget(G)
}
