% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{graphUiInput}
\alias{graphUiInput}
\title{Building blocks for shiny apps and widgets.}
\usage{
graphUiInput()
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Building blocks for shiny apps and widgets.
}

