% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_as_gradget_data}
\alias{igraph_as_gradget_data}
\title{Convert igraph object to gradget data.}
\usage{
igraph_as_gradget_data(graph, nodeSize = 8, nodeColor = "0xcccccc",
  edgeWidth = 5, edgeColor = "0xeeeeee", fontSize = 16,
  fontColor = "#FFFFFF")
}
\arguments{
\item{nodeSize}{Size of nodes (defaults to 8).}

\item{edgeWidth}{Width of edges, an integer (defaults to 5).}

\item{edgeColor}{Color of edges, a hex value (defaults to "0xeeeeee").}

\item{fontSize}{Size of node text, an integer (defaults to 20).}

\item{fontColor}{Color of node text, a hex value (defaults to "#FFFFFF").}
}
\description{
Convert igraph object to gradget data.
}
