% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_as_svg}
\alias{igraph_as_svg}
\title{Turn igraph object into SVG.}
\usage{
igraph_as_svg(graph, radius_min = 5, radius_tf = TRUE,
  edgeColor = "black", fontSize = 8, textOffset = 3, width = 800,
  height = 800)
}
\arguments{
\item{fontSize}{font size of the vertex labels}

\item{textOffset}{where to put text}

\item{width}{the width of the svg}

\item{height}{the height of the svg}

\item{edgeAttributes}{attributes of edges to maintain}

\item{verticeAttributes}{attributes of vertices to maintain}

\item{margin}{margins of the svg}

\item{edgeAttributes}{attributes of edges for tooltips}

\item{verticeAttributes}{attributes of attributes for tooltips}
}
\description{
Turn igraph object into SVG.
}
\examples{
G <- merkel2008
G <- igraph_add_coordinates(G, layout = "kamada.kawai", dim = 2)
G <- igraph_normalize_coordinates(G, width = 800, height = 800, margin = 50)
S <- igraph_as_svg(G)
}
