% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_perform_action}
\alias{igraph_perform_action}
\title{Perform action on igraph object}
\usage{
igraph_perform_action(x)
}
\arguments{
\item{x}{An \code{igraph} object that has been annotated.}
}
\description{
Perform action on igraph object
}
