% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradgets.R
\docType{data}
\name{merkel2008}
\alias{merkel2008}
\title{Cooccurrence (i)graph of Merkel 2008.}
\format{An igraph object}
\usage{
merkel2008
}
\description{
Cooccurrence (i)graph of Merkel 2008.
}
\examples{
Merkel <- Cooccurrences$new(
  partition = merkel2008,
  p_attribute = "word",
  window = 5L,
  drop = terms_to_drop
)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

bt2008 <- partition(
  "GERMAPARL",
  year = 2008, interjection = "FALSE",
  p_attribute = "word", name = "bt2008"
)
terms_to_drop <- c(
  polmineR::punctuation,
  unlist(noise(p_attributes(bt2008, p_attribute = "word")))
)
BT2008 <- Cooccurrences$new(
  partition = bt2008,
  p_attribute = "word",
  window = 5L,
  drop = terms_to_drop
)
BT2008$count()
BT2008$trim(action = "drop", by.id = TRUE)
BT2008$maths()

Merkel$featureSelection(reference = BT2008, included = TRUE)

G <- Merkel$as.igraph(as.undirected = TRUE)
G <- igraph_add_communities(G, method = "fastgreedy", weights = FALSE)
G <- igraph_add_coordinates(G, layout = "kamada.kawai", dim = 3)
G <- rescale(G, -1000, 1000)
}
\keyword{datasets}
