% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{polmineR.graph-package}
\alias{polmineR.graph-package}
\title{How to get started}
\description{
Find the cgi script that is shipped with the page
system.file("cgi", "graph2kwic.R", package="polmineR.graph")
and copy it to /var/FastRWeb/web.R
Add/modify line in /var/FastRWeb/code/rserve.R:
pkgs <- c("XML", "Cairo", "Matrix", "FastRWeb", "polmineR", "pipeR", "markdown")
To start FastRWeb: sudo /var/FastRWeb/code/start
}
\keyword{package}

