% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradgets.R
\docType{package}
\name{polmineR.graph-package}
\alias{polmineR.graph-package}
\alias{polmineR.graph}
\title{How to get started}
\usage{
polmineR.graph()
}
\description{
Find the cgi script that is shipped with the page 
system.file("cgi", "graph2kwic.R", package="polmineR.graph")
and copy it to /var/FastRWeb/web.R
Add/modify line in /var/FastRWeb/code/rserve.R: 
pkgs <- c("XML", "Cairo", "Matrix", "FastRWeb", "polmineR", "pipeR", "markdown")
To start FastRWeb: sudo /var/FastRWeb/code/start
}
\examples{
\dontrun{
library(polmineR)
library(polmineR.graph)
library(three)

merkel2008 <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_year = "2008", speaker_type = "speech", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(merkel2008, pAttribute = "word"))))

Merkel <- Cooccurrences$new(partition = merkel2008, pAttribute = "word", window = 5L, drop = termsToDrop)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

bt2008 <- partition("PLPRBT", speaker_year = "2008", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(bt2008, pAttribute = "word"))))
BT2008 <- Cooccurrences$new(partition = bt2008, pAttribute = "word", window = 5L, drop = termsToDrop)
BT2008$count()
BT2008$trim(action = "drop", by.id = TRUE)
BT2008$maths()

C <- compare(bt17merkelCooc, bt17cooc)

bt17merkelCollTrimmed <- subset(bt17merkelColl, rank_ll <= 250)
iMerkel <- asIgraph(bt17merkelCollTrimmed)
iMerkelComm <- igraph_add_communities(iMerkel, method="fastgreedy", weights=FALSE)
iMerkel3d <- igraph_add_coordinates(iMerkelComm, layout = "kamada.kawai", dim = 3)
iMerkel3d <- three::rescale(iMerkel3d, -400, 400)
t <- polmineR.graph::as.three(iMerkel3d, bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=3, fontOffset=c(x=10,y=10,z=10))
t <- polmineR.graph::as.three(iMerkel3d, type="raycaster", bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=3, fontOffset=c(x=10,y=10,z=10))
t <- as.three(iMerkel3d, type="anaglyph", bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=3, fontOffset=c(x=10,y=10,z=10))
t <- as.three(iMerkel3d, type="disco", bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=3, fontOffset=c(x=10,y=10,z=10))
toView <- three:::store(t, directory="/Users/blaette/Lab/tmp/three")
shiny::runApp("/Users/blaette/Lab/github/polmineR.graph/inst/three")
browseURL(toView["tmpFileJs"])


#' library(polmineR)
library(polmineR.graph)
use("GermaParl")

merkel2008 <- partition(
  "GERMAPARL", speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word", name = "merkel2008"
  )
terms_to_drop <- c(
  polmineR::punctuation,
  unlist(noise(p_attributes(merkel2008, p_attribute = "word")))
  )


}
\dontrun{
merkel2008 <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_year = "2008", speaker_type = "speech", pAttribute = "word", name = "Merkel2008")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(merkel2008, pAttribute = "word"))))

Merkel <- Cooccurrences$new(partition = merkel2008, pAttribute = "word", window = 5L, drop = termsToDrop)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

bt2008 <- partition("PLPRBT", speaker_year = "2008", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(bt2008, pAttribute = "word"))))
BT2008 <- Cooccurrences$new(partition = bt2008, pAttribute = "word", window = 5L, drop = termsToDrop)
BT2008$count()
BT2008$trim(action = "drop", by.id = TRUE)
BT2008$maths()
}
}
\keyword{package}
