% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{graphUiInput}
\alias{cooccurrencesServer}
\alias{cooccurrencesUiInput}
\alias{cooccurrencesUiOutput}
\alias{graphServer}
\alias{graphUiInput}
\alias{graphUiOutput}
\alias{settingsGraphUiInput}
\title{Building blocks for shiny apps and widgets.}
\usage{
graphUiInput()

settingsGraphUiInput()

graphUiOutput()

graphServer(input, output, session)

cooccurrencesUiInput()

cooccurrencesUiOutput()

cooccurrencesServer(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Building blocks for shiny apps and widgets.
}

