% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.R
\name{two}
\alias{two}
\title{Make svg for collocations graph}
\usage{
two(svg)
}
\description{
The resulting svg graph will have clickable nodes and edges.
}
\examples{
\dontrun{

Y <- SVG$new(G)
Y$make()
Y$as_htmlwidget()

Y$browse()
}
}
