% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.svg,collocations-method}
\alias{as.character,svg-method}
\alias{as.svg}
\alias{as.svg,collocations-method}
\alias{as.svg,igraph-method}
\alias{browse}
\alias{browse,svg-method}
\alias{plot,collocations-method}
\alias{plot,igraph-method}
\alias{plot,svg-method}
\alias{print,svg-method}
\alias{show,svg-method}
\title{Make svg for collocations graph}
\usage{
\S4method{as.svg}{collocations}(object, layout = "kamada.kawai",
  verbose = TRUE, ...)

\S4method{as.svg}{igraph}(object, layout = "kamada.kawai", width = 400,
  height = 400, margin = 50, fontSize = 8, textOffset = 5,
  edgeAttributes = "ll", verticeAttributes = "tf", pandocTab = TRUE,
  mc = FALSE, returnXML = FALSE, verbose = TRUE)

\S4method{print}{svg}(x)

\S4method{as.character}{svg}(x)

\S4method{show}{svg}(object)

\S4method{browse}{svg}(object)
}
\arguments{
\item{object}{the collocation object}

\item{layout}{either "kamada.kawai" or "fruchterman.reingold"}

\item{verbose}{whether to be talkative}

\item{...}{parameters that will be passed}

\item{width}{the width of the svg}

\item{height}{the height of the svg}

\item{margin}{margins of the svg}

\item{fontSize}{font size of the vertex labels}

\item{textOffset}{where to put text}

\item{edgeAttributes}{attributes of edges for tooltips}

\item{verticeAttributes}{attributes of attributes for tooltips}

\item{pandocTab}{logical, whether to format tables with pandoc}

\item{mc}{logical, whether to use multicore}

\item{returnXML}{logical, whether to return XML}

\item{x}{a svg object}
}
\description{
A story to be told

The resulting svg graph will have clickable nodes and edges. To use this functionality
you need to install FastRWeb and start it with sudo /var/FastRWeb/code/start
}
\examples{
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_speaker="Angela Merkel", text_type="speech"), tf="word")
bt17merkelColl <- collocations(bt17merkel, pAttribute="word", mc=TRUE)
bt17merkelCollTrimmed <- trim(bt17merkelColl, cutoff=list(ll=10.83))
iMerkel <- asIgraph(bt17merkelCollTrimmed)
iMerkelComm <- enrich(iMerkel, community=list(method="fastgreedy", weights=FALSE))
merkelSvg <- as.svg(iMerkelComm, width=1000, height=1000)
}
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_speaker="Angela Merkel", text_type="speech"), tf="word")
bt17merkelColl <- collocations(bt17merkel, pAttribute="word", mc=TRUE)
bt17merkelCollTrimmed <- trim(bt17merkelColl, cutoff=list(ll=50))
bt17merkelCollTrimmed <- trim(bt17merkelColl, cutoff=list(ll=10.83, collocateWindowFreq=2))
iMerkel <- asIgraph(bt17merkelCollTrimmed)
iMerkel <- enrich(iMerkel, community=list(method="fastgreedy", weights=FALSE))
merkelSvg <- as.svg(iMerkel, width=1000, height=1000)
show(merkelSvg)
}
}

