% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_methods.R, R/svg_class.R
\docType{methods}
\name{asIgraph,cooccurrences-method}
\alias{asIgraph,cooccurrences-method}
\alias{asIgraph,svg-method}
\title{Turn cooccurrences object into igraph}
\usage{
\S4method{asIgraph}{cooccurrences}(x, edgeAttributes = "ll",
  verticeAttributes = NULL, as.undirected = TRUE)

\S4method{asIgraph}{svg}(x, as.undirected = TRUE)
}
\arguments{
\item{x}{an object}

\item{edgeAttributes}{attributes of edges to maintain}

\item{verticeAttributes}{attributes of vertices to maintain}

\item{as.undirected}{logical, whether to turn object into directed graph}
}
\description{
Turn cooccurrences object into igraph
}

