% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granny.R
\name{granny}
\alias{granny}
\title{Graph Annotation with Shiny.}
\usage{
granny(graph)
}
\arguments{
\item{graph}{An \code{igraph} object that is enhanced for visualisation with
the gradget htmlwidget (x, y, and z coordinates, if applicable; additional
edge and node data).}
}
\description{
A shiny gadget for graph annotation with shiny, or granny in short.
}
\examples{
G <- merkel2008
G <- igraph_add_coordinates(G, layout = "kamada.kawai", dim = 3)
G <- igraph_add_communities(G)
G <- rescale(G, -250, 250)

am2008 <- partition(
  "GERMAPARL",
  speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word"
)
V(G)$kwic <- pblapply(V(G)$name, function(n) as.character(kwic(am2008, query = n, verbose = F)))
V(G)$kwic <- sapply(V(G)$kwic, function(x) paste(x, collapse = "<br/>"))
V(G)$kwic <- unlist(V(G)$kwic)

edge_matrix <- igraph::as_edgelist(merkel2008)
q1 <- sprintf('"\%s" []{0,4} "\%s"', edge_matrix[,1], edge_matrix[,2])
q2 <- sprintf('"\%s" []{0,4} "\%s"', edge_matrix[,2], edge_matrix[,1])
E(G)$kwic <- pblapply(
  split(data.frame(q1, q2, stringsAsFactors = F), f = 1L:length(q1)),
  function(q) as.character(kwic(am2008, query = unlist(q), cqp = T, verbose = F))
)
E(G)$kwic <- sapply(E(G)$kwic, function(x) paste(x, collapse = "<br/>"))

granny(G)
}
