% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granny.R
\name{granny}
\alias{granny}
\title{Graph Annotation with Shiny.}
\usage{
granny(graph)
}
\arguments{
\item{graph}{An \code{igraph} object that is enhanced for visualisation with
the gradget htmlwidget (x, y, and z coordinates, if applicable; additional
edge and node data).}
}
\description{
A shiny gadget for graph annotation with shiny, or granny in short.
}
