% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_add_kwic}
\alias{igraph_add_kwic}
\title{Add KWIC lineview to igraph object.}
\usage{
igraph_add_kwic(graph, subcorpus, verbose = TRUE, progress = TRUE)
}
\arguments{
\item{graph}{An igraph object.}

\item{subcorpus}{A partition.}

\item{verbose}{Logical.}

\item{progress}{Logical.}
}
\description{
Add KWIC lineview to igraph object.
}
\examples{
am2008 <- partition(
  "GERMAPARL",
  speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word"
)
}
