% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_as_svg}
\alias{igraph_as_svg}
\title{Turn igraph object into SVG.}
\usage{
igraph_as_svg(x, radius_min = 5, radius_tf = TRUE, edgeColor = "black",
  fontSize = 8, textOffset = 3, width = 800, height = 800)
}
\arguments{
\item{x}{An \code{igraph} object.}

\item{radius_min}{...}

\item{radius_tf}{...}

\item{edgeColor}{...}

\item{fontSize}{font size of the vertex labels}

\item{textOffset}{where to put text}

\item{width}{the width of the svg}

\item{height}{the height of the svg}
}
\description{
Turn igraph object into SVG.
}
\examples{
library(gradgets)
library(magrittr)
library(igraph)
library(shiny)

G <- merkel2008 \%>\%
  igraph_add_coordinates(layout = "kamada.kawai", dim = 2) \%>\%
  igraph_normalize_coordinates(width = 800, height = 800, margin = 50)
S <- igraph_as_svg(G)
w <- svgPanZoom::svgPanZoom(S)

w <- appendContent(
  w,
  shiny::includeScript(path = system.file(package = "gradgets", "js", "svg_js_extensions.js"))
  )
   
}
