% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{addCoordinates}
\alias{addCommunities}
\alias{addCoordinates}
\alias{normalizeCoordinates}
\alias{rescale}
\alias{selectCommunity}
\alias{tokenNeighborhood}
\title{Utility functions to manipulate igraph objects.}
\usage{
addCoordinates(x, layout = c("kamada.kawai", "fruchterman.reingold",
  "kamada.kawai", "spring", "lgl", "graphopt"), dim = 3)

normalizeCoordinates(x, width, height, margin)

rescale(x, min, max)

addCommunities(x, method = "fastgreedy", weights = FALSE)

selectCommunity(x, no)

tokenNeighborhood(x, order, token)
}
\arguments{
\item{x}{igraph object}

\item{layout}{choose from list}

\item{dim}{no of dimensions}

\item{width}{new width}

\item{height}{new height}

\item{margin}{margin}

\item{min}{min}

\item{max}{max}

\item{method}{}

\item{weights}{use weights for community detection}

\item{no}{number of community}

\item{order}{order of neighborhood}

\item{token}{token at center pf neighborhood}

\item{x}{igraph object}
}
\description{
Utility functions to manipulate igraph objects.

Normalize coordinates of igraph object.
}

