% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_add_coordinates}
\alias{igraph_add_coordinates}
\alias{igraph_normalize_coordinates}
\alias{rescale}
\alias{igraph_add_communities}
\alias{selectCommunity}
\alias{as.dgr_graph}
\alias{as.plotly}
\alias{as.networkD3}
\title{Utility functions to manipulate igraph objects.}
\usage{
igraph_add_coordinates(x, layout = c("kamada.kawai", "fruchterman.reingold",
  "kamada.kawai", "spring", "lgl", "graphopt"), dim = 3)

igraph_normalize_coordinates(x, width, height, margin)

rescale(x, min, max)

igraph_add_communities(x, method = "fastgreedy", weights = FALSE)

selectCommunity(x, no)

as.dgr_graph(x)

as.plotly(x)

as.networkD3(x)
}
\arguments{
\item{x}{igraph object}

\item{layout}{choose from list}

\item{dim}{no of dimensions}

\item{width}{new width}

\item{height}{new height}

\item{margin}{margin}

\item{min}{min}

\item{max}{max}

\item{method}{method for coordinate calculation}

\item{weights}{use weights for community detection}

\item{no}{number of community}
}
\description{
Utility functions to manipulate igraph objects.

Normalize coordinates of igraph object.
}
