% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granny.R
\docType{data}
\encoding{UTF-8}
\name{merkel2008}
\alias{merkel2008}
\title{Cooccurrence (i)graph of Merkel 2008.}
\format{An igraph object}
\usage{
merkel2008
}
\description{
Cooccurrence (i)graph of Merkel 2008.
}
\keyword{datasets}
