% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Three.R
\name{three}
\alias{three}
\title{Generate and manage 3d graph.}
\usage{
three(G, type = "base", settings = list(bgColor = "0xcccccc", nodeSize = 5,
  edgeColor = "0xeeeeee", edgeWidth = 5, fontSize = 12, fontColor = "0x000000",
  fontOffset = c(x = 10, y = 10, z = 10)), width = NULL, height = NULL)
}
\arguments{
\item{type}{display mode}

\item{bgColor}{a hex value}

\item{nodeSize}{size of nods}

\item{edgeColor}{a hex value}

\item{edgeWidth}{defaults to 5}

\item{fontColor}{a hex value}

\item{fontSize}{defaults to 20}

\item{fontOffset}{a list}
}
\description{
Generate 3d graph with three.js
}
\examples{
\dontrun{
library(polmineR)
use("GermaParl")

merkel2008 <- partition(
  "GERMAPARL",
  speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word"
)
terms_to_drop <- c(
  polmineR::punctuation,
  unlist(noise(p_attributes(merkel2008, p_attribute = "word")))
)

Merkel <- Cooccurrences$new(partition = merkel2008, p_attribute = "word", window = 5L, drop = terms_to_trop)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

Merkel$dt <- Merkel$dt[1:250]

G <- Merkel$as.igraph()
G <- addCoordinates(G, layout = "kamada.kawai", dim = 3)
G <- addCommunities(G)
G <- rescale(G, -275, 275)

three.settings <- list(fontSize = 18, edgeWidth = 6, edgeColor = "0x666666", bgColor = "0x000000", fontColor <- "0xffffff"
T <- three(G, type "anaglyph", settings = three.settings)
}
}
