% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three.R
\name{three}
\alias{three}
\title{Generate and manage 3d graph.}
\usage{
three(x, bgColor = "0x888888", raycaster = TRUE, anaglyph = FALSE,
  width = NULL, height = NULL)
}
\arguments{
\item{bgColor}{Background color, a hex value (defaults to "0x888888").}
}
\description{
Generate 3d graph with three.js
}
\examples{
\dontrun{
library(polmineR.graph)
library(polmineR)
use("GermaParl")

merkel2008 <- partition(
  "GERMAPARL",
  speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word"
)

terms_to_drop <- c(
  polmineR::punctuation,
  unlist(noise(p_attributes(merkel2008, p_attribute = "word")))
)

Merkel <- Cooccurrences$new(
  partition = merkel2008, p_attribute = "word", window = 5L,
  drop = terms_to_drop
)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

Merkel$dt <- Merkel$dt[1:250]

G <- Merkel$as.igraph()
G <- igraph_add_coordinates(G, layout = "kamada.kawai", dim = 3)
G <- igraph_add_communities(G)
G <- rescale(G, -250, 250)

T <- three(
  igraph_as_gradget_data(G),
  raycaster = TRUE
)
T
}
}
