% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Three.R
\docType{class}
\name{Three}
\alias{Three}
\title{Generate and manage 3d graph.}
\description{
Generate 3d graph with three.js
}
\section{Fields}{

\describe{
\item{\code{type}}{display mode}

\item{\code{bgColor}}{a hex value}

\item{\code{nodeSize}}{size of nods}

\item{\code{edgeColor}}{a hex value}

\item{\code{edgeWidth}}{defaults to 5}

\item{\code{fontColor}}{a hex value}

\item{\code{fontSize}}{defaults to 20}

\item{\code{fontOffset}}{a list}
}}

\section{Methods}{

\describe{
\item{\code{as.json()}}{Turn into json.}

\item{\code{initialize(x, dir = character())}}{Create new Three object.}

\item{\code{make()}}{Fill three slot.}
}}

\examples{
\dontrun{
library(polmineR)
library(rjson)

merkel2008 <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_year = "2008", speaker_type = "speech", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(merkel2008, pAttribute = "word"))))

Merkel <- Cooccurrences$new(partition = merkel2008, pAttribute = "word", window = 5L, drop = termsToDrop)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

Merkel$dt <- Merkel$dt[1:250]

G <- Merkel$as.igraph()
G <- addCoordinates(G, layout = "kamada.kawai", dim = 3)
G <- addCommunities(G)
G <- rescale(G, -275, 275)

T <- Three$new(G, dir = "/Users/blaette/Lab/tmp/three")
T$fontSize <- 18
T$edgeWidth <- 6
T$edgeColor <- "0x666666"
T$bgColor <- "0x000000"
T$fontColor <- "0xffffff"
T$type <- "anaglyph"
T$make()
T$browse()
}
}
