% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Three.R
\docType{class}
\name{Three}
\alias{Three}
\title{Generate and manage 3d graph.}
\description{
Generate 3d graph with three.js
}
\section{Fields}{

\describe{
\item{\code{type}}{display mode}

\item{\code{bgColor}}{a hex value}

\item{\code{nodeSize}}{size of nods}

\item{\code{edgeColor}}{a hex value}

\item{\code{edgeWidth}}{defaults to 5}

\item{\code{fontColor}}{a hex value}

\item{\code{fontSize}}{defaults to 20}

\item{\code{fontOffset}}{a list}
}}
\section{Methods}{

\describe{
\item{\code{as.json()}}{Turn into json.}

\item{\code{initialize(x)}}{Create new Three object.}

\item{\code{make()}}{Fill three slot.}
}}
\examples{
\dontrun{
library(polmineR)
merkel2008 <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_year = "2008", speaker_type = "speech", pAttribute = "word")
termsToDrop <- c(polmineR::punctuation, unlist(noise(pAttributes(merkel2008, pAttribute = "word"))))
Merkel <- Cooccurrences$new(partition = merkel2008, pAttribute = "word", window = 5L, drop = termsToDrop)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

Merkel$dt <- Merkel$dt[1:250]

G <- Merkel$as.igraph()
T <- Three$new(G)

}
}

