% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Three.R
\name{three}
\alias{three}
\title{Generate and manage 3d graph.}
\usage{
three(G = G, type = "base", bgColor = "0x888888", nodeSize = 8,
  edgeColor = "0xeeeeee", edgeWidth = 5, fontSize = 16,
  fontColor = "#FFFFFF", fontOffset = c(x = 10, y = 10, z = 10),
  raycaster = FALSE, anaglyph = FALSE, width = NULL, height = NULL,
  sizing_policy = htmlwidgets::sizingPolicy(padding = 0))
}
\arguments{
\item{type}{display mode}

\item{bgColor}{a hex value}

\item{nodeSize}{size of nods}

\item{edgeColor}{a hex value}

\item{edgeWidth}{defaults to 5}

\item{fontSize}{defaults to 20}

\item{fontColor}{a hex value}

\item{fontOffset}{a list}
}
\description{
Generate 3d graph with three.js
}
\examples{
\dontrun{
library(polmineR.graph)
library(polmineR)
use("GermaParl")

merkel2008 <- partition(
  "GERMAPARL",
  speaker = "Angela Merkel", year = 2008, interjection = FALSE,
  p_attribute = "word"
)

terms_to_drop <- c(
  polmineR::punctuation,
  unlist(noise(p_attributes(merkel2008, p_attribute = "word")))
)

Merkel <- Cooccurrences$new(
  partition = merkel2008, p_attribute = "word", window = 5L,
  drop = terms_to_drop
)
Merkel$count()
Merkel$trim(action = "drop", by.id = TRUE)
Merkel$maths()

Merkel$dt <- Merkel$dt[1:250]

G <- Merkel$as.igraph()
G <- addCoordinates(G, layout = "kamada.kawai", dim = 3)
G <- addCommunities(G)
G <- rescale(G, -250, 250)

T <- three(G, type = "base", raycaster = TRUE)
T
}
}
